requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==24.2.0
blessed==1.20.0
bpython==0.24
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
curtsies==0.4.2
cwcwidth==0.1.9
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
editor==1.6.6
executing==2.0.1
fairscale==0.4.13
fastapi==0.109.2
feedparser==6.0.11
filelock==3.13.1
flake8==7.0.0
flatbuffers==23.5.26
fsspec==2024.2.0
gguf==0.6.0
ghp-import==2.1.0
google-auth==2.28.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
griffe==0.40.1
grpcio==1.60.1
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
identify==2.5.35
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
inquirer==3.2.4
ipykernel==6.29.2
ipython==8.21.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
kubernetes==29.0.0
llama_cpp_python==0.2.45
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-autorefs==0.5.0
mkdocs-material==9.5.10
mkdocs-material-extensions==1.3.1
mkdocstrings==0.24.0
mkdocstrings-python==1.8.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.12.0
opencv-python==4.9.0.80
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
paginate==0.5.6
parso==0.8.3
pathspec==0.12.1
peewee==3.17.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
posthog==3.4.2
pre-commit==3.6.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
Pygments==2.17.2
-e git+ssh://git@github.com/teleprint-me/py.gpt.prompt.git@39b1c4e51047939df98857335da0cc0cbc0813d8#egg=pygptprompt
pymdown-extensions==10.7
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
python-poppler==0.4.1
pyupgrade==3.15.1
pyxdg==0.28
PyYAML==6.0.1
pyyaml_env_tag==0.1
pyzmq==25.1.2
readchar==4.0.5
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
runs==1.2.2
safetensors==0.4.2
selenium==4.18.1
sentencepiece==0.1.99
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenize-rt==5.2.0
tokenizers==0.15.2
torch==2.2.0+cpu
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.0
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xmod==1.8.1
zipp==3.17.0
